moved proxySet to be the main one
